Fix dict comment scrubber; Refactor error maker (breaking)
The function to remove JSON items with commented keys which used
to throw errors in a variety of scenarios was rewritten to work
reliably, but it has not been exhaustively tested.

make_error_message() kwargs were changed. This will break base.py until
a later commit. (In lieu of a parameter named 'while', I settled for 'whilst'...)