Fix asyncio.Queue.get `RuntimeError: Event loop is closed`
The bar's event loop closing on _shutdown() comes as a surprise to the
queue handler while it waits. I'm convinced one has to manually close
its coroutine to prevent it from emitting a RuntimeError. I do that
right before calling _loop.stop(). I think it's a really dunb solution,
and I wish it weren't a problem, but it seems to work.